GAIL-DeepPath/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# scripts/demo.py: 11,12
# scripts/evaluate.py: 10
# scripts/evaluate_att.py: 12
# scripts/evaluate_iter.py: 12
# scripts/evaluate_test.py: 11
Keras == 2.2.0

# scripts/pretrain.py: 16
# scripts/retrain.py: 16
design-nest == 0.1.0

# scripts/pretrain.py: 16
# scripts/retrain.py: 16
django-random-id-generator == 0.1.0

# scripts/demo.py: 6
# scripts/evaluate_att.py: 7
# scripts/evaluate_test.py: 6
# scripts/fact_prediction_eval.py: 6
# scripts/pretrain.py: 10
# scripts/retrain.py: 10
env == 0.1.0

# scripts/evaluate_iter.py: 7
h5py == 2.8.0

# scripts/pretrain.py: 16
# scripts/retrain.py: 16
mee6-image-generator == 1.0.1

# scripts/demo.py: 5
# scripts/demo_sampling.py: 3
# scripts/discriminator.py: 5
# scripts/env.py: 1
# scripts/evaluate.py: 5
# scripts/evaluate_att.py: 5
# scripts/evaluate_iter.py: 6
# scripts/evaluate_test.py: 5
# scripts/fact_prediction_eval.py: 4
# scripts/generator.py: 4
# scripts/pretrain.py: 4
# scripts/retrain.py: 4
# scripts/transE_eval.py: 5
# scripts/transH_eval.py: 5
# scripts/transR_eval.py: 5
# scripts/utils.py: 5
numpy == 1.14.3

# scripts/pretrain.py: 16
# scripts/retrain.py: 16
py-generator == 0.2.1

# scripts/pretrain.py: 16
# scripts/retrain.py: 16
pysyn == 0.0.21

# scripts/pretrain.py: 16
# scripts/retrain.py: 16
rest-cli == 0.0.3

# scripts/demo.py: 9
# scripts/evaluate.py: 8
# scripts/evaluate_att.py: 10
# scripts/evaluate_iter.py: 9
# scripts/evaluate_test.py: 9
sklearn == 0.0

# scripts/demo_sampling.py: 2
# scripts/discriminator.py: 7
# scripts/generator.py: 3
# scripts/networks.py: 1
# scripts/pretrain.py: 3
# scripts/retrain.py: 3
tensorflow_gpu == 1.4.0

# scripts/pretrain.py: 16
# scripts/retrain.py: 16
test-generator == 0.1.2

# scripts/demo.py: 7
# scripts/demo_sampling.py: 8
# scripts/discriminator.py: 14
# scripts/env.py: 3
# scripts/evaluate_att.py: 8
# scripts/evaluate_test.py: 7
# scripts/fact_prediction_eval.py: 7
# scripts/generator.py: 8
# scripts/pretrain.py: 9
# scripts/retrain.py: 9
utils == 0.9.0

# scripts/discriminator.py: 13
# scripts/generator.py: 7
webvirtmgr == 4.8.9
